requirements.txt
serial
psycopg2-binary